mv and small tweaks
