+++ packages/map/package.json
@@ -62,6 +62,7 @@
   },
   "typesVersions": {},
   "devDependencies": {
-    "solid-js": "^1.8.7"
+    "solid-js": "^1.8.7",
+    "solid-transition-group": "^0.2.3"
   }
 }
